File to plot data and added sparse matrix code
Can now plot data on a previously existing file, saving you from having to rerun the entire code each time.

Added code from quantum for sparse matrix as this is next to be benchmarked. I need to write the code to benchmark this,
both in asymptotic time complexity for multiple sparsities and sparsity dependance of running time. I.e the second meaning
for a constant dimension size, what is the effect of dereasing the sparsity from 1 to 0.