Initialise Sparse Matrix by density
To investigate the effect of matrix sparsity (equally, density) on the asymptotic time complexity of
the multiplication algorithm, needed to be able to define a SparseMatrix by a given density, then
be able to use this initialiser when calculating the time complexity of the algorithm.

TODO:

Write code which calculates time complexity for a number of densities, plot the result.
Use to investigate improvements to sparse mm.
Add special cases to the sparsemm algorithm to exploit common features of quantum operators,
i.e diagonal, stream diagonal etc.